backend/requirements.txt
fastapi
uvicorn
ariadne
ariadne-graphql-modules
mangum
boto3
python-jose
requests
